Move code to npmFunctions v1.0.101
