+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "ex-storage",
-  "version": "0.0.12",
+  "version": "0.0.14",
   "private": true,
   "main": "lib/index.js",
   "engines": {
@@ -17,7 +17,7 @@
     "@microsoft/sp-office-ui-fabric-core": "1.11.0",
     "@microsoft/sp-property-pane": "1.11.0",
     "@microsoft/sp-webpart-base": "1.11.0",
-    "@mikezimm/npmfunctions": "^1.0.100",
+    "@mikezimm/npmfunctions": "^1.0.101",
     "@pnp/spfx-controls-react": "3.3.0",
     "@pnp/spfx-property-controls": "^3.2.0",
     "office-ui-fabric-react": "^6.214.0",
