+++ src/webparts/exStorage/components/HelpInfo/Content/About.tsx
@@ -3,6 +3,8 @@ import * as React from 'react';
 //import { IHelpTableRow, IHelpTable, IPageContent, ISinglePageProps } from '../Component/ISinglePageProps';
 import { IHelpTableRow, IHelpTable, IPageContent, ISinglePageProps } from '@mikezimm/npmfunctions/dist/HelpInfo/Component/ISinglePageProps';
 
+export const panelVersionNumber = '1.0.0.14 - 2021-09-28'; //Added to show in panel
+
 export function aboutTable() {
 
     let table : IHelpTable  = {
@@ -11,6 +13,7 @@ export function aboutTable() {
         rows: [],
     };
 
+    table.rows.push( createAboutRow('2021-09-31', '1.0.0.14', `npmFunctions update.` ) );
     table.rows.push( createAboutRow('2021-09-30', '1.0.0.12', `Fix typos in help, Add Tricks logic, improve analytics` ) );
     table.rows.push( createAboutRow('2021-09-29', '1.0.0.11', `Update Banner to show actual webpart specific help, add Analytics` ) );
     table.rows.push( createAboutRow('2021-09-28', '1.0.0.10', `Add Library Dropdown, Improve Summary, Dups, Items, Preview, Media tags` ) );
