+++ src/webparts/exStorage/components/HelpInfo/Content/FuturePlans.tsx
@@ -7,12 +7,12 @@ import { IHelpTableRow, IHelpTable, IPageContent, ISinglePageProps } from '@mike
 
 export function futureContent() {
 
+    return null;
+
     let html1 = <div>
         <h2>Were thinking of making this an extension so it doesn't need to be added to a page!</h2>
     </div>;
 
-    return null;
-    
     return { html1: html1 };
 
 }
