+++ src/webparts/exStorage/components/HelpInfo/banner/Cards/component.module.scss
