+++ src/webparts/exStorage/components/HelpInfo/banner/bannerProps.ts
@@ -1,11 +1,11 @@
+import { IRepoLinks } from '@mikezimm/npmfunctions/dist/HelpInfo/Links/CreateLinks';
 
 export interface IWebpartBannerProps {
 	title: string;
 	style: string;
 	showBanner: boolean;
 	showTricks: boolean;
-	
-	gitHubRepo: any; // replace with IRepoLinks from npmFunctions v0.1.0.3
+	gitHubRepo: IRepoLinks; // replace with IRepoLinks from npmFunctions v0.1.0.3
 	
 }
