+++ src/webparts/exStorage/components/pages/summary/ExAgeSummary.tsx
@@ -8,7 +8,7 @@ import {
 
 } from "office-ui-fabric-react";
 
-import { sortObjectArrayByNumberKey, sortNumberArray } from '@mikezimm/npmfunctions/dist/Services/Arrays/sorting';
+import { sortObjectArrayByChildNumberKey, sortNumberArray } from '@mikezimm/npmfunctions/dist/Services/Arrays/sorting';
 
 import { getStorageItems, batchSize, createBatchData, getSizeLabel } from '../../ExFunctions';
